+++ sequencesender/sequencesender_x1.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"errors"
 	"fmt"
+	ethman "github.com/0xPolygonHermez/zkevm-node/etherman"
 	"github.com/0xPolygonHermez/zkevm-node/etherman/types"
 	"github.com/0xPolygonHermez/zkevm-node/ethtxmanager"
 	"github.com/0xPolygonHermez/zkevm-node/log"
@@ -205,4 +206,92 @@ func (s *SequenceSender) getSequencesToSendX1(ctx context.Context) ([]types.Sequ
 
 	log.Info("not enough time has passed since last batch was virtualized, and the sequence could be bigger")
 	return nil, nil
+}
+
+// handleEstimateGasSendSequenceErr handles an error on the estimate gas. It will return:
+// nil, error: impossible to handle gracefully
+// sequence, nil: handled gracefully. Potentially manipulating the sequences
+// nil, nil: a situation that requires waiting
+func (s *SequenceSender) handleEstimateGasSendSequenceErr(
+	ctx context.Context,
+	sequences []types.Sequence,
+	currentBatchNumToSequence uint64,
+	err error,
+) ([]types.Sequence, error) {
+	// Insufficient allowance
+	if errors.Is(err, ethman.ErrInsufficientAllowance) {
+		return nil, err
+	}
+	if isDataForEthTxTooBig(err) {
+		// Remove the latest item and send the sequences
+		log.Infof(
+			"Done building sequences, selected batches to %d. Batch %d caused the L1 tx to be too big",
+			currentBatchNumToSequence-1, currentBatchNumToSequence,
+		)
+		sequences = sequences[:len(sequences)-1]
+		return sequences, nil
+	}
+
+	// while estimating gas a new block is not created and the POE SC may return
+	// an error regarding timestamp verification, this must be handled
+	if errors.Is(err, ethman.ErrTimestampMustBeInsideRange) {
+		// query the sc about the value of its lastTimestamp variable
+		lastTimestamp, err := s.etherman.GetLastBatchTimestamp()
+		if err != nil {
+			return nil, err
+		}
+		// check POE SC lastTimestamp against sequences' one
+		for _, seq := range sequences {
+			if seq.Timestamp < int64(lastTimestamp) {
+				// TODO: gracefully handle this situation by creating an L2 reorg
+				log.Fatalf("sequence timestamp %d is < POE SC lastTimestamp %d", seq.Timestamp, lastTimestamp)
+			}
+			lastTimestamp = uint64(seq.Timestamp)
+		}
+		blockTimestamp, err := s.etherman.GetLatestBlockTimestamp(ctx)
+		if err != nil {
+			log.Error("error getting block timestamp: ", err)
+		}
+		log.Debugf("block.timestamp: %d is smaller than seq.Timestamp: %d. A new block must be mined in L1 before the gas can be estimated.", blockTimestamp, sequences[0].Timestamp)
+		return nil, nil
+	}
+
+	// Unknown error
+	if len(sequences) == 1 {
+		// TODO: gracefully handle this situation by creating an L2 reorg
+		log.Errorf(
+			"Error when estimating gas for BatchNum %d (alone in the sequences): %v",
+			currentBatchNumToSequence, err,
+		)
+	}
+	// Remove the latest item and send the sequences
+	log.Infof(
+		"Done building sequences, selected batches to %d. Batch %d excluded due to unknown error: %v",
+		currentBatchNumToSequence, currentBatchNumToSequence+1, err,
+	)
+	sequences = sequences[:len(sequences)-1]
+
+	return sequences, nil
+}
+
+func isDataForEthTxTooBig(err error) bool {
+	return errors.Is(err, ethman.ErrGasRequiredExceedsAllowance) ||
+		errors.Is(err, ErrOversizedData) ||
+		errors.Is(err, ethman.ErrContentLengthTooLarge)
+}
+
+func (s *SequenceSender) isValidium() bool {
+	if !s.cfg.UseValidium {
+		return false
+	}
+
+	committee, err := s.etherman.GetCurrentDataCommittee()
+	if err != nil {
+		return false
+	}
+
+	if len(committee.Members) <= 0 {
+		return false
+	}
+	return true
 }
